refactor(dashboard): split the dimension editor into one hook per table
